Module nodle_parachain::service

Service and ServiceFactory implementation. Specialized wrapper over substrate service.

Functions

new_partial

Starts a ServiceBuilder for a full service.

parachain_build_import_queue

Build the import queue for the parachain runtime.

start_node_impl 🔒

start_parachain_node

Start a parachain node.

warn_if_slow_hardware 🔒

Checks that the hardware meets the requirements and print a warning otherwise.

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Balance

Balance of an account.

Block

Block type.

BlockNumber

An index to a block.

Hash

A hash of some data used by the chain.

Header

Header type.

HostFunctions 🔒

Nonce

Nonce of a transaction in the chain.

ParachainBackend 🔒

ParachainBlockImport 🔒

ParachainClient 🔒