Function nodle_parachain::service::new_partial

pub fn new_partial<RuntimeApi, Executor, BIQ>(
    config: &Configuration, 
    build_import_queue: BIQ
) -> Result<PartialComponents<TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>, TFullBackend<Block>, (), DefaultImportQueue<Block, TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>>, FullPool<Block, TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>>, (Option<Telemetry>, Option<TelemetryWorkerHandle>)>, Error> where
    RuntimeApi: ConstructRuntimeApi<Block, TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>> + Send + Sync + 'static,
    RuntimeApi::RuntimeApi: TaggedTransactionQueue<Block> + Metadata<Block> + SessionKeys<Block> + ApiExt<Block, StateBackend = StateBackendFor<TFullBackend<Block>, Block>> + OffchainWorkerApi<Block> + BlockBuilder<Block>,
    StateBackendFor<TFullBackend<Block>, Block>: StateBackend<BlakeTwo256>,
    Executor: NativeExecutionDispatch + 'static,
    BIQ: FnOnce(Arc<TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>>, &Configuration, Option<TelemetryHandle>, &TaskManager) -> Result<DefaultImportQueue<Block, TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>>, Error>, 

Starts a ServiceBuilder for a full service.

Use this macro if you don’t actually need the full service, but just the builder in order to be able to perform chain operations.