Function nodle_parachain::service::start_node_impl

async fn start_node_impl<RuntimeApi, Executor, RB, BIQ, BIC>(
    parachain_config: Configuration,
    polkadot_config: Configuration,
    collator_options: CollatorOptions,
    id: ParaId,
    _rpc_ext_builder: RB,
    build_import_queue: BIQ,
    build_consensus: BIC,
    hwbench: Option<HwBench>
) -> Result<(TaskManager, Arc<TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>>)>where
    RuntimeApi: ConstructRuntimeApi<Block, TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>> + Send + Sync + 'static,
    RuntimeApi::RuntimeApi: TaggedTransactionQueue<Block> + Metadata<Block> + SessionKeys<Block> + ApiExt<Block, StateBackend = StateBackendFor<TFullBackend<Block>, Block>> + OffchainWorkerApi<Block> + BlockBuilder<Block> + CollectCollationInfo<Block> + TransactionPaymentRuntimeApi<Block, Balance> + ContractsRuntimeApi<Block, AccountId, Balance, BlockNumber, Hash> + AccountNonceApi<Block, AccountId, Nonce>,
    StateBackendFor<TFullBackend<Block>, Block>: StateBackend<BlakeTwo256>,
    Executor: NativeExecutionDispatch + 'static,
    RB: Fn(Arc<TFullClient<Block, RuntimeApi, Executor>>) -> Result<RpcModule<()>, Error> + Send + 'static,
    BIQ: FnOnce(Arc<TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>>, &Configuration, Option<TelemetryHandle>, &TaskManager) -> Result<DefaultImportQueue<Block, TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>>, Error> + 'static,
    BIC: FnOnce(Arc<TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>>, Option<&Registry>, Option<TelemetryHandle>, &TaskManager, Arc<dyn RelayChainInterface>, Arc<FullPool<Block, TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>>>, Arc<NetworkService<Block, Hash>>, SyncCryptoStorePtr, bool) -> Result<Box<dyn ParachainConsensus<Block>>, Error>,

Start a node with the given parachain Configuration and relay chain Configuration.

This is the actual implementation that is abstract over the executor and the runtime api.