Function nodle_parachain::service::new_partial

pub fn new_partial(
    config: &Configuration
) -> Result<PartialComponents<TFullClient<Block, RuntimeApi, WasmExecutor<(SubstrateHostFunctions, HostFunctions)>>, TFullBackend<Block>, (), DefaultImportQueue<Block>, FullPool<Block, TFullClient<Block, RuntimeApi, WasmExecutor<(SubstrateHostFunctions, HostFunctions)>>>, (ParachainBlockImport<Block, Arc<TFullClient<Block, RuntimeApi, WasmExecutor<(SubstrateHostFunctions, HostFunctions)>>>, TFullBackend<Block>>, Option<Telemetry>, Option<TelemetryWorkerHandle>)>, Error>

Starts a ServiceBuilder for a full service.

Use this macro if you don’t actually need the full service, but just the builder in order to be able to perform chain operations.